Format change and create export sbml util module
