+++ mira/sources/sbml/utils.py
@@ -79,7 +79,6 @@ def get_model_annotations(sbml_model, *, converter, logger):
     if not ann_xml:
         return None
     et = etree.fromstring(ann_xml)
-
     annot_structure = {
         "publications": "bqmodel:isDescribedBy",
         "diseases": "bqbiol:is",
@@ -365,37 +364,7 @@ def process_unit_definition(unit_definition):
     return full_unit_expr
 
 
-def get_uri(curie: str) -> Optional[str]:
-    """Convert a curie to a URI, prioritizing the miriam format."""
-    return bioregistry.get_iri(
-        curie, priority=["miriam", "bioregistry", "default"]
-    )
 
-
-def create_biological_cv_term(
-    resource, qualifier_predicate
-) -> Optional[libsbml.CVTerm]:
-    uri_resource = get_uri(resource)
-    if not uri_resource:
-        return None
-    term = libsbml.CVTerm()
-    term.setQualifierType(libsbml.BIOLOGICAL_QUALIFIER)
-    term.setBiologicalQualifierType(qualifier_predicate)
-    term.addResource(uri_resource)
-    return term
-
-
-def create_model_cv_term(
-    resource, qualifier_predicate
-) -> Optional[libsbml.CVTerm]:
-    uri_resource = get_uri(resource)
-    if not uri_resource:
-        return None
-    term = libsbml.CVTerm()
-    term.setQualifierType(libsbml.MODEL_QUALIFIER)
-    term.setModelQualifierType(qualifier_predicate)
-    term.addResource(uri_resource)
-    return term
 
 
 unit_symbol_mappings = {
